0.0.2

+++ main.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"os"
 	"sort"
+	"strings"
 	"syscall"
 
 	clipboard "github.com/atotto/clipboard"
 	goutils "github.com/simonski/goutils"
-
 	terminal "golang.org/x/crypto/ssh/terminal"
 )
 
@@ -20,7 +20,7 @@ func main() {
 	} else if isInfo(command) {
 		DoInfo(cli)
 	} else if isVerify(command) {
-		DoVerify(cli, false)
+		DoVerify(cli, true)
 	} else if isVersion(command) {
 		DoVersion(cli)
 	} else if isClear(command) {
@@ -28,11 +28,15 @@ func main() {
 	} else if isList(command) {
 		DoList(cli)
 	} else if isPut(command, cli) {
-		// DoVerify(cli, true)
-		DoPut(cli)
+		ok := DoVerify(cli, false)
+		if ok {
+			DoPut(cli)
+		}
 	} else if isGet(command, cli) {
-		// DoVerify(cli, true)
-		DoGet(cli)
+		ok := DoVerify(cli, false)
+		if ok {
+			DoGet(cli)
+		}
 	} else if isDelete(command) {
 		DoDelete(cli)
 	} else if command != "" {
@@ -88,13 +92,13 @@ func LoadDB() *CrypticDB {
 
 // DoVerify performs verification of ~/.Crypticfile, encryption/decryption using
 // specified keys
-func DoVerify(cli *goutils.CLI, failOnError bool) {
+func DoVerify(cli *goutils.CLI, printFailuresToStdOut bool) bool {
 	overallValid := true
 	filename := goutils.GetEnvOrDefault(CRYPTIC_FILE, "~/.Crypticfile")
 	publicKey := goutils.GetEnvOrDefault(CRYPTIC_PUBLIC_KEY, "~/.ssh/id_rsa.pem")
 	privateKey := goutils.GetEnvOrDefault(CRYPTIC_PRIVATE_KEY, "~/.ssh/id_rsa")
 
-	filenameExists := goutils.FileExists(goutils.EvaluateFilename(filename))
+	// filenameExists := goutils.FileExists(goutils.EvaluateFilename(filename))
 	publicKeyExists := goutils.FileExists(goutils.EvaluateFilename(publicKey))
 	privateKeyExists := goutils.FileExists(goutils.EvaluateFilename(privateKey))
 
@@ -103,15 +107,15 @@ func DoVerify(cli *goutils.CLI, failOnError bool) {
 	messages = append(messages, fmt.Sprintf("%v    %v\n", CRYPTIC_PUBLIC_KEY, publicKey))
 	messages = append(messages, fmt.Sprintf("%v   %v\n", CRYPTIC_PRIVATE_KEY, privateKey))
 
-	overallValid = filenameExists && publicKeyExists && privateKeyExists
-	if !filenameExists {
-		line := fmt.Sprintf("Crypticfile '%v' does not exist.\n", filename)
-		messages = append(messages, line)
-		overallValid = false
-	} else {
-		// line := fmt.Sprintf("Crypticfile '%v' exists.\n", filename)
-		// messages = append(messages, line)
-	}
+	// overallValid = filenameExists && publicKeyExists && privateKeyExists
+	// if !filenameExists {
+	// 	line := fmt.Sprintf("Crypticfile '%v' does not exist.\n", filename)
+	// 	messages = append(messages, line)
+	// 	overallValid = false
+	// } else {
+	// 	// line := fmt.Sprintf("Crypticfile '%v' exists.\n", filename)
+	// 	// messages = append(messages, line)
+	// }
 
 	if !publicKeyExists {
 		line := fmt.Sprintf("Public key '%v' does not exist.\n", publicKey)
@@ -153,15 +157,21 @@ func DoVerify(cli *goutils.CLI, failOnError bool) {
 		messages = append(messages, line)
 	}
 
-	for _, line := range messages {
-		fmt.Printf(line)
+	if printFailuresToStdOut {
+		for _, line := range messages {
+			fmt.Printf(line)
+		}
 	}
+
 	if overallValid {
-		fmt.Printf("cryptic verify : OK.\n")
+		if printFailuresToStdOut {
+			fmt.Printf("cryptic verify : OK.\n")
+		}
 	} else {
 		// fmt.Printf("cryptic verify: NOT OK.\n")
 	}
 
+	return overallValid
 	// fmt.Printf("%v =%v, exists=%v\n", CRYPTIC_ENCRYPTION_ENABLED, privateKeyExists)
 }
 
@@ -198,6 +208,10 @@ func DoPut(cli *goutils.CLI) {
 	db := LoadDB()
 	command := cli.GetCommand()
 	key := cli.GetStringOrDie(command)
+	if len(key) > 25 {
+		fmt.Printf("Error, key must be <= 25 characters.\n")
+		os.Exit(1)
+	}
 	description := cli.GetStringOrDefault("-d", "")
 	password := ""
 	if cli.IndexOf("-value") > -1 {
@@ -237,17 +251,58 @@ func DoList(cli *goutils.CLI) {
 			keys = append(keys, key)
 		}
 		sort.Strings(keys)
-		fmt.Printf("--------------------------------------------------------------------------\n")
-		fmt.Printf("| Key       | Description          | Last Updated      | Created         |\n")
-		fmt.Printf("--------------------------------------------------------------------------\n")
+
+		key_width := 50
+		desc_width := 50
+		date_width := 25
+		width := key_width + desc_width + (2 * date_width) + 4
+
+		line := strings.Repeat("-", width) + "\n"
+		fmt.Printf(line)
+		header := fmt.Sprintf("| Key%v| Description%v| Updated%v| Created%v|\n", strings.Repeat(" ", key_width-len("Key")-1), strings.Repeat(" ", desc_width-len("Description")-1), strings.Repeat(" ", date_width-len("Updated")-1), strings.Repeat(" ", date_width-len("Created")-1))
+		fmt.Printf(header)
+		fmt.Printf(line)
+		max_key_length := key_width - 5
+		max_description_length := desc_width - 5
 		for _, key := range keys {
-			// fmt.Printf("%v\n", key)
 			entry := data.Entries[key]
-			createdStr := entry.Created.Format("January 2, 2006")
-			updatedStr := entry.LastUpdated.Format("January 2, 2006")
-			fmt.Printf("| %-10v| %-20v | %v | %v |\n", key, entry.Description, updatedStr, createdStr)
+			created := entry.Created.Format("January 2, 2006")
+			updated := entry.LastUpdated.Format("January 2, 2006")
+			desc := entry.Description
+
+			if len(key) > max_key_length {
+				key = key[0:max_key_length] + "..."
+			}
+
+			if desc == "" {
+				desc = "No description."
+			}
+			if len(desc) > max_description_length {
+				desc = desc[0:max_description_length] + "..."
+			}
+
+			keyExtra := key_width - 1 - len(key)
+			descExtra := desc_width - 1 - len(desc)
+			updatedExtra := date_width - 1 - len(updated)
+			createdExtra := date_width - 1 - len(created)
+
+			if keyExtra < 0 {
+				keyExtra = 0
+			}
+			if descExtra < 0 {
+				descExtra = 0
+			}
+			if updatedExtra < 0 {
+				updatedExtra = 0
+			}
+			if createdExtra < 0 {
+				createdExtra = 0
+			}
+
+			entry_line := fmt.Sprintf("| %v%v| %v%v| %v%v| %v%v|\n", key, strings.Repeat(" ", keyExtra), desc, strings.Repeat(" ", descExtra), updated, strings.Repeat(" ", updatedExtra), created, strings.Repeat(" ", createdExtra))
+			fmt.Printf(entry_line)
 		}
-		fmt.Printf("--------------------------------------------------------------------------\n")
+		fmt.Printf(line)
 	}
 }
 
@@ -259,6 +314,11 @@ func DoDelete(cli *goutils.CLI) {
 		fmt.Printf("%v\n", USAGE)
 	}
 	db := LoadDB()
+	_, exists := db.Get(key)
+	if !exists {
+		fmt.Printf("Error, '%v' does not exist.\n", key)
+		os.Exit(1)
+	}
 	db.Delete(key)
 	db.Save()
